src/test/java/org/huebert/iotfsdb/partition/LongPartitionTest.java
package org.huebert.iotfsdb.partition;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import java.nio.ByteBuffer;

import static org.assertj.core.api.Assertions.assertThat;

public class LongPartitionTest {

    private static final int SIZE = 10;

    private static final int TYPE_SIZE = Long.BYTES;

    private static final int NUM_BYTES = SIZE * TYPE_SIZE;

    private final PartitionAdapter adapter = new LongPartition();

    private ByteBuffer buffer;

    @BeforeEach
    public void beforeEach() {
        buffer = ByteBuffer.allocate(NUM_BYTES);
        for (int i = 0; i < SIZE; i++) {
            buffer.asLongBuffer().put(i, i);
        }
    }

    @Test
    public void testGetTypeSize() {
        assertThat(adapter.getTypeSize()).isEqualTo(TYPE_SIZE);
    }

    @Test
    public void testPut() {
        adapter.put(buffer, 1, 1.234);
        assertThat(buffer.asLongBuffer().get(1)).isEqualTo(1);
        adapter.put(buffer, 1, null);
        assertThat(buffer.asLongBuffer().get(1)).isEqualTo(Long.MIN_VALUE);
    }

    @Test
    public void testStream() {
        assertThat(adapter.getStream(buffer, 0, SIZE).mapToInt(Number::intValue).sum()).isEqualTo(45);
    }

}
